[crm]: Typecast to unit64_t to avoid divide by 0 during overflow (#1550)
Fix orchagent crash in case of overflow resulting in divide by 0

+++ orchagent/crmorch.cpp
@@ -541,7 +541,7 @@ void CrmOrch::checkCrmThresholds()
 
             if (cnt.usedCounter != 0)
             {
-                percentageUtil = (cnt.usedCounter * 100) / (cnt.usedCounter + cnt.availableCounter);
+                percentageUtil = (uint32_t)((cnt.usedCounter * 100) / (uint64_t)(cnt.usedCounter + cnt.availableCounter));
             }
 
             switch (res.thresholdType)
